# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Fridge AI Backend API
  version: 1.0.0
  description: OpenAPI spec generated from current Spring controllers.
servers:
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Get all users
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserList'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Success
...
  /api/v1/users/{userId}/meals:
    get:
      tags: [Planner]
      summary: Get meal plans by user
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMealPlanList'
    post:
      tags: [Planner]
      summary: Create meal plan
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealPlanRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMealPlan'
  /api/v1/users/{userId}/meals/{mealId}:
    get:
      tags: [Planner]
      summary: Get meal plan by ID
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/MealIdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMealPlan'
    put:
      tags: [Planner]
      summary: Update meal plan
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/MealIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealPlanRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMealPlan'
    delete:
      tags: [Planner]
      summary: Delete meal plan
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/MealIdPath'
      responses:
        '200':
          description: Success
  /api/v1/recipe-shares:
    get:
      tags: [Social]
      summary: Get all recipe shares
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecipeShareList'
    post:
      tags: [Social]
      summary: Create recipe share
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeShareRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecipeShare'
  /api/v1/recipe-shares/{id}:
    get:
      tags: [Social]
      summary: Get recipe share by ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecipeShare'
    put:
      tags: [Social]
      summary: Update recipe share
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeShareRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecipeShare'
    delete:
      tags: [Social]
      summary: Delete recipe share
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Success
  /api/v1/recipe-shares/user/{userId}:
    get:
      tags: [Social]
      summary: Get recipe shares by user
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecipeShareList'
  /api/v1/recipe-shares/{id}/like:
    post:
      tags: [Social]
      summary: Like recipe share
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Success
...
  /api/v1/analytics/events/user/{userId}:
    get:
      tags: [Analytics]
      summary: Get analytics events by user
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAnalyticsEventList'
...
components:
  parameters:
...
    XUserIdHeader:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    MealIdPath:
      name: mealId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    UserRequest:
      type: object
      additionalProperties: true
    UserResponse:
      type: object
      additionalProperties: true
...
    MealPlanRequest:
      type: object
      additionalProperties: true
    MealPlanResponse:
      type: object
      additionalProperties: true
    RecipeShareRequest:
      type: object
      additionalProperties: true
    RecipeShareResponse:
      type: object
      additionalProperties: true
...
    AnalyticsEventResponse:
      type: object
      additionalProperties: true
...
    ApiResponseUser:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserResponse'
    ApiResponseUserList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
...
    ApiResponseMealPlan:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/MealPlanResponse'
    ApiResponseMealPlanList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/MealPlanResponse'
    ApiResponseRecipeShare:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/RecipeShareResponse'
    ApiResponseRecipeShareList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/RecipeShareResponse'
...
    ApiResponseAnalyticsEventList:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsEventResponse'